Update 03.1_quickstart-creation.md
Lots of small suggestions and typo fixes. One thing I'd question is the inclusion of rdfs:range and rdfs:domain though (as they are conceptually complex and abstract, and I rarely use them at all (as they tend to be very restrictive (hence Schema.org defining the much looser 'schema:rangeIncludes' and 'schema:domainIncludes'))).
Overall I think this is really good, but my sense is that it's still too 'advanced' (and long-winded) for a basic introduction to RDF and Turtle. I think it could be split into two maybe, i.e. a really minimal 'intro', and then a 'slightly more advanced' section.