Update 05.1_quickstart-artifacts.md
Just some quick updates. This page still needs a fair bit of work (as does the LIT VocabTerm interface!) - but how do the explicit line numbers work (i.e. 'line 16', 'line 20' and 'line 35'), for instance, what if we add a new line of text at the start of this page - do we have to manually update these line numbers...!?