First step re-implementing updated design
This implementation re-uses Bulma to get a more consistent design
and DOM structure. In doing so, it implements the latest designs
by Lisa.

It also introduces management of the footer through
_data/footer.yml.

The redesigned homepage is in reasonable shape, but I still need to
receive a few assets to be able to complete the work:

- The background image for the white part of the landing page.
- The illustration for "Like peas in a Pod"
- The illustration for "Solid is a standard"

Additionally, we might still want to implement the stretching of
the hero background to also cover the menu, but only on the
landing page.