Attempt to satisfy HTMLProofer
We're getting errors that links to /for-developers/ are invalid,
presumably because a `.html` file doesn't get converted to a URL
without a trailing slash, whereas that _was_ the case for
Markdown files.

It still remains to be seen whether GitHub Pages also allows for
the trailing slash to be left off.