Re-add the spell-checking dictionary
This is the dictionary that, at the time of writing, is used in the
staging website, while the production version does not even have
the spellchecker enabled yet. However, since Travis is apparently
looking at the staging branch when determining its configuration,
it is looking for this dictionary file even when spellchecking has
not yet been set up in this branch.