Update
As discussed in the Friday 15th Editors meeting it would be helpful to have a one-stop-shop where 1) people who are interested in contributing to the spec can find out how 2) people who are impacted by the spec can know how to stay up to date without joining all the calls and chat forums. Hopefully, the one-stop-shop will also help increase the number of the proposal. 

This pull request needs links to W3C wiki with 1) meetings 2) spec updates when that exists