Solid FAQ
There are a few questions on the forum and in response to talks on Solid around privacy and Solid so I propose we try and summarise the essence of the questions in an FAQ along with an answer that addresses the key points. 

In this pull request, there's a very rough jotted down version of an FAQ to kick off the draft. 

In particular, I would like to invite the specification editors to refine this copy.