Create Solid World Invites Directly on Site
Eventbrite is making unusual recommendations to other events and community members are confused thinking that Solid is promoting these external events. Also, community members are not finding the invite possibly due to the need for many clicks. 

As a solution, I propose putting the Solid World Invite directly on solidproject.org. 

This also allows for collaborative editing. 

Things that need to happen to this pull request: 
1) formatting to make it look good on solidproject.org
2) put the .md file somewhere logical in github.com/solid/solidproject.org