Create apps-table.html
Solid is a platform, built using the existing web. It gives every user a choice about where data is stored, which specific people and groups can access select elements, and which apps you use. 
Solid changes the current model where users have to hand over personal data to digital giants in exchange for perceived value. 
Solid is a technology, like the Web, but a new level of standard which adds to the existing protocols to make it more powerful, particularly to empower individuals at home and at work.r