Adding the Vimeo Link
I added the vimeo link for the Jan SW. BUT In addition, I would like to redo this web page to better the usability. I have been practicing my code in here and it seems to work when I preview the changes. If you want I can put in a PR of the changes I would like to make! It might take me a day or two.