add reference to open spec
As a follow on to [PR 161](https://github.com/solid/solidproject.org/pull/671) - requesting a note to be added to this page near the end to encourage developers to reference the standards that Solid uses. This way if the language of their choice is not listed; they still may be able to interact with Solid if their language has frameworks already in place to work with HTTP, OIDC, etc.

I don't know that I'm the greatest wordsmith, so happy to accept other changes here.